Move PartitionedOutputOperator to repartition folder
