+++ presto-main/src/main/java/com/facebook/presto/operator/OperatorInfo.java
@@ -15,6 +15,7 @@ package com.facebook.presto.operator;
 
 import com.facebook.presto.operator.TableWriterOperator.TableWriterInfo;
 import com.facebook.presto.operator.exchange.LocalExchangeBufferInfo;
+import com.facebook.presto.operator.repartition.PartitionedOutputInfo;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
