+++ presto-main/src/main/java/com/facebook/presto/operator/repartition/PartitionedOutputInfo.java
@@ -11,8 +11,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.operator;
+package com.facebook.presto.operator.repartition;
 
+import com.facebook.presto.operator.OperatorInfo;
 import com.facebook.presto.util.Mergeable;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
